Move browser-independent extension code into src/common
This makes it possible to use
https://github.com/ccnmtl/mediathread-collect
as the central repo for common extension code.

See:
http://www.columbia.edu/~njn2118/journal/2015/11/13.html
