+++ ChangeLog
@@ -1,3 +1,7 @@
+2015-11-13  Nik Nyby  <nnyby@columbia.edu>
+
+	* Moved browser-independent extension code into src/common.
+
 2015-11-09  Nik Nyby  <nnyby@columbia.edu>
 
 	* Fixed bug where video thumbnails could be missing when
