+++ src/common/asset-handler.js
