+++ src/common/collect.js
