+++ src/common/host-handler.js
